# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReineiraOS Modules API",
    "version": "0.1.0",
    "description": "Backend API for ReineiraOS venture applications"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/withdrawals": {
      "post": {
        "operationId": "createWithdrawal",
        "tags": [
          "Withdrawal"
        ],
        "summary": "Create a new withdrawal",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "escrow_ids": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "minItems": 1,
                    "maxItems": 100
                  },
                  "destination_chain": {
                    "type": "string",
                    "enum": [
                      "BASE",
                      "ETH",
                      "POLYGON"
                    ]
                  },
                  "recipient_address": {
                    "type": "string",
                    "pattern": "^0x[a-fA-F0-9]{40}$"
                  }
                },
                "required": [
                  "escrow_ids",
                  "destination_chain",
                  "recipient_address"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Withdrawal created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "public_id": {
                      "type": "string"
                    },
                    "calls": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "contract_address": {
                            "type": "string"
                          },
                          "abi_function_signature": {
                            "type": "string"
                          },
                          "abi_parameters": {
                            "type": "object",
                            "additionalProperties": {}
                          }
                        },
                        "required": [
                          "contract_address",
                          "abi_function_signature",
                          "abi_parameters"
                        ]
                      }
                    },
                    "status": {
                      "type": "string"
                    },
                    "estimated_amount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "public_id",
                    "calls",
                    "status",
                    "estimated_amount"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "statusCode"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "statusCode"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listWithdrawals",
        "tags": [
          "Withdrawal"
        ],
        "summary": "List withdrawals for current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated withdrawal list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "public_id": {
                            "type": "string"
                          },
                          "escrow_ids": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            }
                          },
                          "destination_chain": {
                            "type": "string"
                          },
                          "recipient_address": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "estimated_amount": {
                            "type": "number"
                          },
                          "actual_amount": {
                            "type": "number"
                          },
                          "fee": {
                            "type": "number"
                          },
                          "redeem_tx_hash": {
                            "type": "string"
                          },
                          "bridge_tx_hash": {
                            "type": "string"
                          },
                          "destination_tx_hash": {
                            "type": "string"
                          },
                          "error_message": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          },
                          "completed_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "public_id",
                          "escrow_ids",
                          "destination_chain",
                          "recipient_address",
                          "status",
                          "estimated_amount",
                          "created_at",
                          "updated_at"
                        ]
                      }
                    },
                    "continuation_token": {
                      "type": "string"
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "limit": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "items",
                    "has_more",
                    "limit"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "error",
                    "message",
                    "statusCode"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}